--- src/components/Autocomplete/index.js
@@ -1,2 +1,0 @@
-export { default as Autocomplete } from './Autocomplete';
-export { default as ItemTag } from './ItemTag';
