+++ src/components/Autosuggest/ItemTag/__tests__/ItemTag.spec.js
