+++ src/components/Autosuggest/__mocks__/suggestions.js
