+++ src/components/Autosuggest/index.js
@@ -1,0 +1,2 @@
+export { default as Autosuggest } from './Autosuggest';
+export { default as ItemTag } from './ItemTag';
